+++ apps/frontend/src/index.tsx
@@ -14,6 +14,7 @@ import { store } from "~/libs/modules/store/store.js";
 import { AccessManagement } from "~/pages/access-management/access-management.jsx";
 import { Analytics } from "~/pages/analytics/analytics.jsx";
 import { Auth } from "~/pages/auth/auth.jsx";
+import { Contributors } from "~/pages/contributors/contributors.jsx";
 import { NoAccess } from "~/pages/no-access/no-access.jsx";
 import { NotFound } from "~/pages/not-found/not-found.jsx";
 import { Profile } from "~/pages/profile/profile.jsx";
@@ -32,7 +33,10 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 							{
 								element: (
 									<ProtectedRoute
-										routePermissions={[PermissionKey.VIEW_ALL_PROJECTS]}
+										routePermissions={[
+											PermissionKey.VIEW_ALL_PROJECTS,
+											PermissionKey.MANAGE_ALL_PROJECTS,
+										]}
 									>
 										<Projects />
 									</ProtectedRoute>
@@ -56,6 +60,14 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 									</ProtectedRoute>
 								),
 								path: AppRoute.PROFILE,
+							},
+							{
+								element: (
+									<ProtectedRoute>
+										<Contributors />
+									</ProtectedRoute>
+								),
+								path: AppRoute.CONTRIBUTORS,
 							},
 							{
 								element: (
@@ -86,11 +98,26 @@ createRoot(document.querySelector("#root") as HTMLElement).render(
 								path: AppRoute.SIGN_UP,
 							},
 							{
-								element: <Project />,
+								element: (
+									<ProtectedRoute
+										routePermissions={[
+											PermissionKey.VIEW_ALL_PROJECTS,
+											PermissionKey.MANAGE_ALL_PROJECTS,
+										]}
+									>
+										<Project />
+									</ProtectedRoute>
+								),
 								path: AppRoute.PROJECT,
 							},
 							{
-								element: <ProjectAccessManagement />,
+								element: (
+									<ProtectedRoute
+										routePermissions={[PermissionKey.MANAGE_ALL_PROJECTS]}
+									>
+										<ProjectAccessManagement />
+									</ProtectedRoute>
+								),
 								path: AppRoute.PROJECT_ACCESS_MANAGEMENT,
 							},
 						],