+++ apps/frontend/src/pages/project/libs/types/types.ts
@@ -1,0 +1,1 @@
+export { type ContributorGetAllItemResponseDto } from "@git-fit/shared";