+++ packages/shared/src/libs/enums/permission-key.enum.ts
@@ -1,4 +1,5 @@
 const PermissionKey = {
+	MANAGE_ALL_PROJECTS: "manage_all_projects",
 	MANAGE_USER_ACCESS: "manage_user_access",
 	VIEW_ALL_PROJECTS: "view_all_projects",
 } as const;