+++ packages/shared/src/modules/contributors/libs/enums/contributors-api-path.enum.ts
@@ -1,0 +1,5 @@
+const ContributorsApiPath = {
+	ROOT: "/",
+} as const;
+
+export { ContributorsApiPath };